rename LangchainAgent to ReActAgent and AIAgent to Agent
- Rename ai-agent.ts to agent.ts, AIAgent interface to Agent, AiMessage to AgentMessage
- Rename langchain-agent.ts to react-agent.ts, LangchainAgent to ReActAgent
- Remove AnyToolSignature alias, inline ToolSignature<any> at usage site
- Update all imports across server, services, and tests
